+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for merging coverage reports.
 
+### Changed
+
+- Exclude case statements in `BeginEndRequiredRule`.
+
 ## [0.34.2] - 2022-03-03
 
 ### Added