+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support for MSBuild conditionals
 - Exclude the `Tools` folder from standard library indexation.
 - Exclude case statements in `BeginEndRequiredRule`.
+- Exclude `.dpk` files in `EmptyUnitRule`.
 
 ## [0.34.2] - 2022-03-03