+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rewrite project file parsing from scratch:
   - Support for property interpolation
   - Support for MSBuild conditionals
+- Exclude the `Tools` folder from standard library indexation.
 - Exclude case statements in `BeginEndRequiredRule`.
 
 ## [0.34.2] - 2022-03-03